Added pytest flag to retain S3 output as reference
Added pytest flag to retain S3 output if desired to use as reference for regression testing

+++ tests/test_MIRI.py
@@ -23,7 +23,7 @@ from eureka.S5_lightcurve_fitting import s5_fit as s5
 from eureka.S6_planet_spectra import s6_spectra as s6
 
 
-def test_MIRI(capsys):
+def test_MIRI(capsys, keep_s3_output):
     s2_installed = 'eureka.S2_calibrations.s2_calibrate' in sys.modules
     if not s2_installed:
         with capsys.disabled():
@@ -169,7 +169,8 @@ def test_MIRI(capsys):
         #           f"Stage1{os.sep}S1_*")
         os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}"
                   f"Stage2{os.sep}S2_*")
-    os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}Stage3")
+    if not keep_s3_output:
+        os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}Stage3")
     os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}Stage4")
     os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}Stage5")
     os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}MIRI{os.sep}Stage6")
