+++ tests/test_Photometry_NIRCam.py
@@ -13,7 +13,7 @@ from eureka.S4_generate_lightcurves import s4_genLC as s4
 from eureka.S5_lightcurve_fitting import s5_fit as s5
 
 
-def test_NIRCamPhotometry(capsys):
+def test_NIRCamPhotometry(capsys, keep_s3_output):
     with capsys.disabled():
         # is able to display any message without failing a test
         # useful to leave messages for future users who run the tests
@@ -73,12 +73,13 @@ def test_NIRCamPhotometry(capsys):
     assert np.array_equal(s5_meta.citations, s5_cites)
 
     # remove temporary files
-    os.system(f"rm -r data{os.sep}Photometry{os.sep}NIRCam{os.sep}Stage3")
+    if not keep_s3_output:
+        os.system(f"rm -r data{os.sep}Photometry{os.sep}NIRCam{os.sep}Stage3")
     os.system(f"rm -r data{os.sep}Photometry{os.sep}NIRCam{os.sep}Stage4")
     os.system(f"rm -r data{os.sep}Photometry{os.sep}NIRCam{os.sep}Stage5")
 
 
-def test_NIRCamPhotometry_hex(capsys):
+def test_NIRCamPhotometry_hex(capsys, keep_s3_output):
     # tests hexagonal photometry aperture
     with capsys.disabled():
         # is able to display any message without failing a test
@@ -112,4 +113,5 @@ def test_NIRCamPhotometry_hex(capsys):
     assert np.array_equal(s3_meta.citations, s3_cites)
 
     # remove temporary files
-    os.system(f"rm -r data{os.sep}Photometry{os.sep}NIRCam{os.sep}Stage3")
+    if not keep_s3_output:
+        os.system(f"rm -r data{os.sep}Photometry{os.sep}NIRCam{os.sep}Stage3")
